requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
pydantic
assemblyai
murf
google-generativeai